hello_world/data/requirements.txt
psycopg2-binary==2.9.5
sqlalchemy
